server/requirements.txt
fastapi
httpx
uvicorn
requests
python-dotenv